⏪️ Fjerner DatoInput komponent
Det var ikke så lett å generalisere datosetter siden det er forskjellig logikk for fom og tom
