+++ packages/frontend/routes/saksbilde/sykepengegrunnlag/inntekt/Refusjon/useRefusjonFormField.tsx
@@ -3,7 +3,7 @@ import { useFieldArray, useFormContext } from 'react-hook-form';
 import { Kildetype } from '@io/graphql';
 import { Refusjonsopplysning } from '@io/http';
 
-interface RefusjonFormValues {
+export interface RefusjonFormValues {
     name: string;
     refusjonsopplysninger: { fom: string; tom?: Maybe<string>; beløp: number; kilde: string }[];
 }
